Add basic testing for GameControllerManager
There's currently no testing of the java code for game-controller. This cl allows testing of the constructor and some trivial methods in GameControllerManager, and enables additions to test further methods.

Also cleaned up the build.gradle as it looks like the buildTest variable is unnecessary as you can already declare if the dependencies are for the tests or not

Bug: b/429403537
Test: ./gradlew games-controller:connectedAndroidTest
Change-Id: Ie9286910ee2d1b0f3b8e541aaf61cb5df0309272